+++ backend/src/role/models/user-roles.model.ts
@@ -1,5 +1,5 @@
 import { Column, DataType, ForeignKey, Model, Table } from 'sequelize-typescript';
-import { UserModel } from '../../users/models/user.model';
+import { UserModel } from '../../user/models/user.model';
 import { RoleModel } from './role.model';
 
 @Table({ tableName: 'user_roles', createdAt: false, updatedAt: false })
@@ -8,11 +8,12 @@ export class UserRolesModel extends Model<UserRolesModel> {
     @Column({ type: DataType.INTEGER, unique: true, autoIncrement: true, primaryKey: true })
     id: number;
 
+    @ForeignKey(() => RoleModel)
+    @Column({type: DataType.INTEGER})
+    roleId: number;
+
     @ForeignKey(() => UserModel)
     @Column({ type: DataType.INTEGER })
     userId: number;
 
-    @ForeignKey(() => RoleModel)
-    @Column({type: DataType.INTEGER})
-    roleId: number;
 }